# Optimize logger init performance by using module-level constants